+++ CHANGELOG.md
@@ -19,6 +19,8 @@ how a consumer would use the library (e.g. adding unit tests, updating documenta
 
 ### Added
 
+- Debug logs for restore progress during the method call `py42.archive.stream_from_back()`.
+
 - `py42.constants.SortDirection` constants `DESC` and `ASC`.
 
 - `sdk.detectionlists.departing_employee.DepartingEmployeeFilters` constants `OPEN`,
@@ -82,6 +84,8 @@ how a consumer would use the library (e.g. adding unit tests, updating documenta
 
 ### Changed
 
+- `py42.archive.stream_from_backup()` now calculates file sizes and accepts a `file_size_calc_timeout` parameter.
+- Parameter `file_path` on `sdk.archive.stream_from_backup()` renamed to `file_paths` and can now take a list of file paths to restore.
 - `py42.detectionlists.departing_employee.add()` now raises `Py42UserAlreadyAddedError` when the user is already on the list.
 - `py42.detectionlists.high_risk_employee.add()` now raises `Py42UserAlreadyAddedError` when the user already on the list.
 - `py42.legalhold.add_to_matter()` now raises `Py42UserAlreadyAddedError` when the user is already on the matter.